プロジェクト設定を変更

+++ main.cpp
@@ -25,7 +25,7 @@ int32_t main(int32_t argc, char* argv[])
 	}
 
 
-	for (uint32_t i = 0; i < argc; i++)
+	for (size_t i = 0; i < argc; i++)
 	{
 		cout << argv[i] << endl;
 	}
